requirements.txt
<<<<<<< HEAD
python-dotenv
pytz
psutil
supabase
gotrue
httpx
postgrest
realtime
storage3
supafunc
# picamera2  # Replaced with OpenCV for better reliability
ffmpeg-python
Flask
Pillow
numpy
requests
boto3
fastapi
uvicorn
opencv-python
speedtest-cli
=======
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP requests
requests==2.31.0

# Database and cloud
supabase==2.0.2
boto3==1.34.0
botocore==1.34.0

# System utilities
psutil==5.9.6
python-dotenv==1.0.0

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Template engine
jinja2==3.1.2

# Computer vision and video processing
opencv-python-headless==4.8.1.78
numpy>=2.0.0,<2.3.0

# Date and time utilities
python-dateutil==2.8.2
pytz==2023.3

# Additional dependencies for enhanced functionality
portalocker==2.7.0
dataclasses-json==0.6.1

# Video processing dependencies
av>=15.0.0
piexif==1.1.3
pillow==10.1.0
python-prctl==1.8.1
videodev2==0.0.1
libarchive-c==2.9

# JSON schema validation
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
referencing==0.31.0
rpds-py==0.13.0
typing-extensions>=4.12.0

# Progress bars and utilities
tqdm==4.66.1
attrs==23.1.0

# Camera and image processing
# Note: picamera2 is installed via apt system packages
PiDNG==4.0.9
>>>>>>> 7f4d06de69b6359ae09f590d27a614501e93bf81
